Fixed alot of the UI and in-game scripts
I still need to look over the boundaries and get new music for the win/death screen, but i put some filler for now. Sydney, look over it and see if you can improve some of the coloring and stuff. If someone wants to add a particle system for the dash, go for it.